+++ bukkit/src/main/java/de/maxhenkel/voicechat/plugins/impl/PositionImpl.java
@@ -3,6 +3,8 @@ package de.maxhenkel.voicechat.plugins.impl;
 import de.maxhenkel.voicechat.api.Position;
 import org.bukkit.Location;
 
+import java.util.Objects;
+
 public class PositionImpl implements Position {
 
     private final Location position;
@@ -32,5 +34,22 @@ public class PositionImpl implements Position {
 
     public Location getPosition() {
         return position;
+    }
+
+    @Override
+    public boolean equals(Object object) {
+        if (this == object) {
+            return true;
+        }
+        if (object == null || getClass() != object.getClass()) {
+            return false;
+        }
+        PositionImpl position1 = (PositionImpl) object;
+        return Objects.equals(position, position1.position);
+    }
+
+    @Override
+    public int hashCode() {
+        return position != null ? position.hashCode() : 0;
     }
 }