+++ common/src/main/java/de/maxhenkel/voicechat/plugins/impl/ServerLevelImpl.java
@@ -2,6 +2,8 @@ package de.maxhenkel.voicechat.plugins.impl;
 
 import de.maxhenkel.voicechat.api.ServerLevel;
 
+import java.util.Objects;
+
 public class ServerLevelImpl implements ServerLevel {
 
     private final net.minecraft.server.level.ServerLevel serverLevel;
@@ -13,5 +15,22 @@ public class ServerLevelImpl implements ServerLevel {
     @Override
     public Object getServerLevel() {
         return serverLevel;
+    }
+
+    @Override
+    public boolean equals(Object object) {
+        if (this == object) {
+            return true;
+        }
+        if (object == null || getClass() != object.getClass()) {
+            return false;
+        }
+        ServerLevelImpl that = (ServerLevelImpl) object;
+        return Objects.equals(serverLevel, that.serverLevel);
+    }
+
+    @Override
+    public int hashCode() {
+        return serverLevel != null ? serverLevel.hashCode() : 0;
     }
 }